[jest-expo] Move automatic typescript path mapping to `src` to avoid dependency cycle
