+++ packages/jest-expo/config/index.js
@@ -4,13 +4,11 @@ const {
   getAndroidPreset,
   getNodePreset,
 } = require('./getPlatformPreset');
-const { withTypescriptMapping } = require('./withTypescriptMapping');
 const { getWatchPlugins, withWatchPlugins } = require('./withWatchPlugins');
 
 module.exports = {
   getWatchPlugins,
   withWatchPlugins,
-  withTypescriptMapping,
   getWebPreset,
   getIOSPreset,
   getAndroidPreset,
