+++ packages/jest-expo/jest-preset.js
@@ -5,7 +5,7 @@ const isEqual = require('lodash/isEqual');
 // Derive the Expo Jest preset from the React Native one
 const jestPreset = cloneDeep(require('react-native/jest-preset'));
 
-const { withTypescriptMapping } = require('./config');
+const { withTypescriptMapping } = require('./src/preset/withTypescriptMapping');
 
 // transform
 if (!jestPreset.transform) {
