+++ packages/jest-expo/src/preset/withTypescriptMapping.js
